Use bottom tabs to nav mobile sandbox
