--- src/pages/Default.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import Vaults from '../components/Vaults';
-
-function DefaultPage(){
-	return <Vaults />;
-}
-
-export default DefaultPage;
